Add files via upload
Datalink:https://archive.ics.uci.edu/ml/machine-learning-databases/00391/
The Data set come from UCI machine learning repository site. In original data,  5 types of hand postures from 12 users were recorded using unlabeled markers attached to fingers of a glove in a motion capture environment. A rigid pattern of markers on the back of the glove was used to establish a local coordinate system for the hand, and 11 other markers were attached to the thumb and fingers of the glove.
The data set can be used for multiple purpose. In the report, for classification purpose, the features of different hand postures were used to recognize different kind of posture types. 
After updating the data set, one of the classes with small amount of  information has been deleted. 
There were 15822 observations in the dataset, observations were classified into 4 classes. Class size was showed in the following table: